backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.28.1
slowapi==0.1.9
python-dotenv==1.0.1
itsdangerous==2.2.0
aiofiles==24.1.0
rank-bm25==0.2.2
feedparser==6.0.11
beautifulsoup4==4.12.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
anyio==4.7.0
